% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{accept_shared_directory}
\alias{accept_shared_directory}
\title{Accepts a directory sharing request that was sent from the directory owner account}
\usage{
accept_shared_directory(SharedDirectoryId)
}
\arguments{
\item{SharedDirectoryId}{[required] Identifier of the shared directory in the directory consumer account. This identifier is different for each directory owner account.}
}
\description{
Accepts a directory sharing request that was sent from the directory owner account.
}
\section{Accepted Parameters}{
\preformatted{accept_shared_directory(
  SharedDirectoryId = "string"
)
}
}

