% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{connect_directory}
\alias{connect_directory}
\title{Creates an AD Connector to connect to an on-premises directory}
\usage{
connect_directory(Name, Password, Size, ConnectSettings, ShortName = NULL,
  Description = NULL)
}
\arguments{
\item{Name}{The fully qualified name of the on-premises directory, such as \code{corp.example.com}.}

\item{Password}{The password for the on-premises user account.}

\item{Size}{The size of the directory.}

\item{ConnectSettings}{A DirectoryConnectSettings object that contains additional information for the operation.}

\item{ShortName}{The NetBIOS name of the on-premises directory, such as \code{CORP}.}

\item{Description}{A textual description for the directory.}
}
\description{
Creates an AD Connector to connect to an on-premises directory.
}
\details{
Before you call \code{ConnectDirectory}, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the \code{ConnectDirectory} operation, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/UsingWithDS_IAM_ResourcePermissions.html}{AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference}.
}
\examples{

}
