% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{create_conditional_forwarder}
\alias{create_conditional_forwarder}
\title{Creates a conditional forwarder associated with your AWS directory}
\usage{
create_conditional_forwarder(DirectoryId, RemoteDomainName, DnsIpAddrs)
}
\arguments{
\item{DirectoryId}{[required] The directory ID of the AWS directory for which you are creating the conditional forwarder.}

\item{RemoteDomainName}{[required] The fully qualified domain name (FQDN) of the remote domain with which you will set up a trust relationship.}

\item{DnsIpAddrs}{[required] The IP addresses of the remote DNS server associated with RemoteDomainName.}
}
\description{
Creates a conditional forwarder associated with your AWS directory. Conditional forwarders are required in order to set up a trust relationship with another domain. The conditional forwarder points to the trusted domain.
}
\section{Accepted Parameters}{
\preformatted{create_conditional_forwarder(
  DirectoryId = "string",
  RemoteDomainName = "string",
  DnsIpAddrs = list(
    "string"
  )
)
}
}

