% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{create_directory}
\alias{create_directory}
\title{Creates a Simple AD directory}
\usage{
create_directory(Name, Password, Size, ShortName = NULL, Description = NULL,
  VpcSettings = NULL)
}
\arguments{
\item{Name}{The fully qualified name for the directory, such as \code{corp.example.com}.}

\item{Password}{The password for the directory administrator. The directory creation process creates a directory administrator account with the user name \code{Administrator} and this password.}

\item{Size}{The size of the directory.}

\item{ShortName}{The short name of the directory, such as \code{CORP}.}

\item{Description}{A textual description for the directory.}

\item{VpcSettings}{A DirectoryVpcSettings object that contains additional information for the operation.}
}
\description{
Creates a Simple AD directory.
}
\details{
Before you call \code{CreateDirectory}, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the \code{CreateDirectory} operation, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/UsingWithDS_IAM_ResourcePermissions.html}{AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference}.
}
\examples{

}
