% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{create_microsoft_ad}
\alias{create_microsoft_ad}
\title{Creates an AWS Managed Microsoft AD directory}
\usage{
create_microsoft_ad(Name, ShortName = NULL, Password,
  Description = NULL, VpcSettings, Edition = NULL)
}
\arguments{
\item{Name}{[required] The fully qualified domain name for the directory, such as \code{corp.example.com}. This name will resolve inside your VPC only. It does not need to be publicly resolvable.}

\item{ShortName}{The NetBIOS name for your domain. A short identifier for your domain, such as \code{CORP}. If you don't specify a NetBIOS name, it will default to the first part of your directory DNS. For example, \code{CORP} for the directory DNS \code{corp.example.com}.}

\item{Password}{[required] The password for the default administrative user named \code{Admin}.}

\item{Description}{A textual description for the directory. This label will appear on the AWS console \code{Directory Details} page after the directory is created.}

\item{VpcSettings}{[required] Contains VPC information for the CreateDirectory or CreateMicrosoftAD operation.}

\item{Edition}{AWS Managed Microsoft AD is available in two editions: Standard and Enterprise. Enterprise is the default.}
}
\description{
Creates an AWS Managed Microsoft AD directory.
}
\details{
Before you call \emph{CreateMicrosoftAD}, ensure that all of the required permissions have been explicitly granted through a policy. For details about what permissions are required to run the \emph{CreateMicrosoftAD} operation, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/UsingWithDS_IAM_ResourcePermissions.html}{AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference}.
}
\section{Accepted Parameters}{
\preformatted{create_microsoft_ad(
  Name = "string",
  ShortName = "string",
  Password = "string",
  Description = "string",
  VpcSettings = list(
    VpcId = "string",
    SubnetIds = list(
      "string"
    )
  ),
  Edition = "Enterprise"|"Standard"
)
}
}

