% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{create_snapshot}
\alias{create_snapshot}
\title{Creates a snapshot of a Simple AD or Microsoft AD directory in the AWS cloud}
\usage{
create_snapshot(DirectoryId, Name = NULL)
}
\arguments{
\item{DirectoryId}{The identifier of the directory of which to take a snapshot.}

\item{Name}{The descriptive name to apply to the snapshot.}
}
\description{
Creates a snapshot of a Simple AD or Microsoft AD directory in the AWS cloud.
}
\details{
You cannot take snapshots of AD Connector directories.
}
\examples{

}
