% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{delete_conditional_forwarder}
\alias{delete_conditional_forwarder}
\title{Deletes a conditional forwarder that has been set up for your AWS directory}
\usage{
delete_conditional_forwarder(DirectoryId, RemoteDomainName)
}
\arguments{
\item{DirectoryId}{[required] The directory ID for which you are deleting the conditional forwarder.}

\item{RemoteDomainName}{[required] The fully qualified domain name (FQDN) of the remote domain with which you are deleting the conditional forwarder.}
}
\description{
Deletes a conditional forwarder that has been set up for your AWS directory.
}
\section{Accepted Parameters}{
\preformatted{delete_conditional_forwarder(
  DirectoryId = "string",
  RemoteDomainName = "string"
)
}
}

