% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{delete_log_subscription}
\alias{delete_log_subscription}
\title{Deletes the specified log subscription}
\usage{
delete_log_subscription(
  DirectoryId = "string"
)
}
\arguments{
\item{DirectoryId}{[required] Identifier (ID) of the directory whose log subscription you want to delete.}
}
\description{
Deletes the specified log subscription.
}
\examples{

}
