% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{delete_snapshot}
\alias{delete_snapshot}
\title{Deletes a directory snapshot}
\usage{
delete_snapshot(SnapshotId)
}
\arguments{
\item{SnapshotId}{The identifier of the directory snapshot to be deleted.}
}
\description{
Deletes a directory snapshot.
}
\examples{

}
