% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{delete_trust}
\alias{delete_trust}
\title{Deletes an existing trust relationship between your AWS Managed Microsoft AD directory and an external domain}
\usage{
delete_trust(TrustId, DeleteAssociatedConditionalForwarder = NULL)
}
\arguments{
\item{TrustId}{The Trust ID of the trust relationship to be deleted.}

\item{DeleteAssociatedConditionalForwarder}{Delete a conditional forwarder as part of a DeleteTrustRequest.}
}
\description{
Deletes an existing trust relationship between your AWS Managed Microsoft AD directory and an external domain.
}
\examples{

}
