% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{deregister_event_topic}
\alias{deregister_event_topic}
\title{Removes the specified directory as a publisher to the specified SNS topic}
\usage{
deregister_event_topic(DirectoryId, TopicName)
}
\arguments{
\item{DirectoryId}{The Directory ID to remove as a publisher. This directory will no longer send messages to the specified SNS topic.}

\item{TopicName}{The name of the SNS topic from which to remove the directory as a publisher.}
}
\description{
Removes the specified directory as a publisher to the specified SNS topic.
}
\examples{

}
