% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{describe_conditional_forwarders}
\alias{describe_conditional_forwarders}
\title{Obtains information about the conditional forwarders for this account}
\usage{
describe_conditional_forwarders(DirectoryId, RemoteDomainNames = NULL)
}
\arguments{
\item{DirectoryId}{The directory ID for which to get the list of associated conditional forwarders.}

\item{RemoteDomainNames}{The fully qualified domain names (FQDN) of the remote domains for which to get the list of associated conditional forwarders. If this member is null, all conditional forwarders are returned.}
}
\description{
Obtains information about the conditional forwarders for this account.
}
\details{
If no input parameters are provided for RemoteDomainNames, this request describes all conditional forwarders for the specified directory ID.
}
\examples{

}
