% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{describe_domain_controllers}
\alias{describe_domain_controllers}
\title{Provides information about any domain controllers in your directory}
\usage{
describe_domain_controllers(DirectoryId, DomainControllerIds = NULL,
  NextToken = NULL, Limit = NULL)
}
\arguments{
\item{DirectoryId}{Identifier of the directory for which to retrieve the domain controller information.}

\item{DomainControllerIds}{A list of identifiers for the domain controllers whose information will be provided.}

\item{NextToken}{The \emph{DescribeDomainControllers.NextToken} value from a previous call to DescribeDomainControllers. Pass null if this is the first call.}

\item{Limit}{The maximum number of items to return.}
}
\description{
Provides information about any domain controllers in your directory.
}
\examples{

}
