% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{describe_event_topics}
\alias{describe_event_topics}
\title{Obtains information about which SNS topics receive status messages from the specified directory}
\usage{
describe_event_topics(
  DirectoryId = "string",
  TopicNames = list(
    "string"
  )
)
}
\arguments{
\item{DirectoryId}{The Directory ID for which to get the list of associated SNS topics. If this member is null, associations for all Directory IDs are returned.}

\item{TopicNames}{A list of SNS topic names for which to obtain the information. If this member is null, all associations for the specified Directory ID are returned.

An empty list results in an \code{InvalidParameterException} being thrown.}
}
\description{
Obtains information about which SNS topics receive status messages from the specified directory.
}
\details{
If no input parameters are provided, such as DirectoryId or TopicName, this request describes all of the associations in the account.
}
\examples{

}
