% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{describe_shared_directories}
\alias{describe_shared_directories}
\title{Returns the shared directories in your account}
\usage{
describe_shared_directories(OwnerDirectoryId, SharedDirectoryIds = NULL,
  NextToken = NULL, Limit = NULL)
}
\arguments{
\item{OwnerDirectoryId}{Returns the identifier of the directory in the directory owner account.}

\item{SharedDirectoryIds}{A list of identifiers of all shared directories in your account.}

\item{NextToken}{The \code{DescribeSharedDirectoriesResult.NextToken} value from a previous call to DescribeSharedDirectories. Pass null if this is the first call.}

\item{Limit}{The number of shared directories to return in the response object.}
}
\description{
Returns the shared directories in your account.
}
\examples{

}
