% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{disable_radius}
\alias{disable_radius}
\title{Disables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector or Microsoft AD directory}
\usage{
disable_radius(
  DirectoryId = "string"
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory for which to disable MFA.}
}
\description{
Disables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector or Microsoft AD directory.
}
\examples{

}
