% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{get_directory_limits}
\alias{get_directory_limits}
\title{Obtains directory limit information for the current region}
\usage{
get_directory_limits()
}
\description{
Obtains directory limit information for the current region.
}
\section{Accepted Parameters}{
\preformatted{get_directory_limits()
}
}

