% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{list_schema_extensions}
\alias{list_schema_extensions}
\title{Lists all schema extensions applied to a Microsoft AD Directory}
\usage{
list_schema_extensions(
  DirectoryId = "string",
  NextToken = "string",
  Limit = 123
)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory from which to retrieve the schema extension information.}

\item{NextToken}{The \code{ListSchemaExtensions.NextToken} value from a previous call to \code{ListSchemaExtensions}. Pass null if this is the first call.}

\item{Limit}{The maximum number of items to return.}
}
\description{
Lists all schema extensions applied to a Microsoft AD Directory.
}
\examples{

}
