% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists all tags on a directory}
\usage{
list_tags_for_resource(ResourceId, NextToken = NULL, Limit = NULL)
}
\arguments{
\item{ResourceId}{[required] Identifier (ID) of the directory for which you want to retrieve tags.}

\item{NextToken}{Reserved for future use.}

\item{Limit}{Reserved for future use.}
}
\description{
Lists all tags on a directory.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  ResourceId = "string",
  NextToken = "string",
  Limit = 123
)
}
}

