% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{reject_shared_directory}
\alias{reject_shared_directory}
\title{Rejects a directory sharing request that was sent from the directory owner account}
\usage{
reject_shared_directory(
  SharedDirectoryId = "string"
)
}
\arguments{
\item{SharedDirectoryId}{[required] Identifier of the shared directory in the directory consumer account. This identifier is different for each directory owner account.}
}
\description{
Rejects a directory sharing request that was sent from the directory owner account.
}
\examples{

}
