% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{remove_ip_routes}
\alias{remove_ip_routes}
\title{Removes IP address blocks from a directory}
\usage{
remove_ip_routes(DirectoryId, CidrIps)
}
\arguments{
\item{DirectoryId}{Identifier (ID) of the directory from which you want to remove the IP addresses.}

\item{CidrIps}{IP address blocks that you want to remove.}
}
\description{
Removes IP address blocks from a directory.
}
\examples{

}
