% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{remove_tags_from_resource}
\alias{remove_tags_from_resource}
\title{Removes tags from a directory}
\usage{
remove_tags_from_resource(ResourceId, TagKeys)
}
\arguments{
\item{ResourceId}{Identifier (ID) of the directory from which to remove the tag.}

\item{TagKeys}{The tag key (name) of the tag to be removed.}
}
\description{
Removes tags from a directory.
}
\examples{

}
