% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{reset_user_password}
\alias{reset_user_password}
\title{Resets the password for any user in your AWS Managed Microsoft AD or Simple AD directory}
\usage{
reset_user_password(DirectoryId, UserName, NewPassword)
}
\arguments{
\item{DirectoryId}{Identifier of the AWS Managed Microsoft AD or Simple AD directory in which the user resides.}

\item{UserName}{The user name of the user whose password will be reset.}

\item{NewPassword}{The new password that will be reset.}
}
\description{
Resets the password for any user in your AWS Managed Microsoft AD or Simple AD directory.
}
\examples{

}
