% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{start_schema_extension}
\alias{start_schema_extension}
\title{Applies a schema extension to a Microsoft AD directory}
\usage{
start_schema_extension(DirectoryId, CreateSnapshotBeforeSchemaExtension,
  LdifContent, Description)
}
\arguments{
\item{DirectoryId}{The identifier of the directory for which the schema extension will be applied to.}

\item{CreateSnapshotBeforeSchemaExtension}{If true, creates a snapshot of the directory before applying the schema extension.}

\item{LdifContent}{The LDIF file represented as a string. To construct the LdifContent string, precede each line as it would be formatted in an ldif file with \n. See the example request below for more details. The file size can be no larger than 1MB.}

\item{Description}{A description of the schema extension.}
}
\description{
Applies a schema extension to a Microsoft AD directory.
}
\examples{

}
