% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{update_conditional_forwarder}
\alias{update_conditional_forwarder}
\title{Updates a conditional forwarder that has been set up for your AWS directory}
\usage{
update_conditional_forwarder(DirectoryId, RemoteDomainName, DnsIpAddrs)
}
\arguments{
\item{DirectoryId}{[required] The directory ID of the AWS directory for which to update the conditional forwarder.}

\item{RemoteDomainName}{[required] The fully qualified domain name (FQDN) of the remote domain with which you will set up a trust relationship.}

\item{DnsIpAddrs}{[required] The updated IP addresses of the remote DNS server associated with the conditional forwarder.}
}
\description{
Updates a conditional forwarder that has been set up for your AWS directory.
}
\section{Accepted Parameters}{
\preformatted{update_conditional_forwarder(
  DirectoryId = "string",
  RemoteDomainName = "string",
  DnsIpAddrs = list(
    "string"
  )
)
}
}

