% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{update_trust}
\alias{update_trust}
\title{Updates the trust that has been set up between your AWS Managed Microsoft AD directory and an on-premises Active Directory}
\usage{
update_trust(TrustId, SelectiveAuth = NULL)
}
\arguments{
\item{TrustId}{Identifier of the trust relationship.}

\item{SelectiveAuth}{Updates selective authentication for the trust.}
}
\description{
Updates the trust that has been set up between your AWS Managed Microsoft AD directory and an on-premises Active Directory.
}
\examples{

}
