% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{add_tags_to_resource}
\alias{add_tags_to_resource}
\title{Adds or overwrites one or more tags for the specified directory}
\usage{
add_tags_to_resource(
  ResourceId = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
\arguments{
\item{ResourceId}{[required] Identifier (ID) for the directory to which to add the tag.}

\item{Tags}{[required] The tags to be assigned to the directory.}
}
\description{
Adds or overwrites one or more tags for the specified directory. Each directory can have a maximum of 50 tags. Each tag consists of a key and optional value. Tag keys must be unique to each resource.
}
\examples{

}
