% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{enable_radius}
\alias{enable_radius}
\title{Enables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector or Microsoft AD directory}
\usage{
enable_radius(DirectoryId, RadiusSettings)
}
\arguments{
\item{DirectoryId}{The identifier of the directory for which to enable MFA.}

\item{RadiusSettings}{A RadiusSettings object that contains information about the RADIUS server.}
}
\description{
Enables multi-factor authentication (MFA) with the Remote Authentication Dial In User Service (RADIUS) server for an AD Connector or Microsoft AD directory.
}
\examples{

}
