% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{get_snapshot_limits}
\alias{get_snapshot_limits}
\title{Obtains the manual snapshot limits for a directory}
\usage{
get_snapshot_limits(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] Contains the identifier of the directory to obtain the limits for.}
}
\description{
Obtains the manual snapshot limits for a directory.
}
\section{Accepted Parameters}{
\preformatted{get_snapshot_limits(
  DirectoryId = "string"
)
}
}

