% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.directoryservice_operations.R
\name{unshare_directory}
\alias{unshare_directory}
\title{Stops the directory sharing between the directory owner and consumer accounts}
\usage{
unshare_directory(DirectoryId, UnshareTarget)
}
\arguments{
\item{DirectoryId}{The identifier of the AWS Managed Microsoft AD directory that you want to stop sharing.}

\item{UnshareTarget}{Identifier for the directory consumer account with whom the directory has to be unshared.}
}
\description{
Stops the directory sharing between the directory owner and consumer accounts.
}
\examples{

}
